+++ src/main/java/com/antman/extensionmarket42/services/ExtensionServices/base/ExtensionService.java
@@ -3,6 +3,6 @@ package com.antman.extensionmarket42.services.ExtensionServices.base;
 import com.antman.extensionmarket42.models.Extension;
 
 public interface ExtensionService {
-    Extension getById(int id);
+    Extension getById(long id);
     Extension save(Extension extension);
 }
